moved the class to util function
